requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.11,<3.13
psycopg2-binary>=2.8
django-cors-headers